+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/GoldenRecordUpdateServices.kt
@@ -461,6 +461,7 @@ class GoldenRecordUpdateChunkService(
     private fun RelationType.toGateModel() =
         when(this){
             RelationType.IsAlternativeHeadquarterFor -> SharableRelationType.IsAlternativeHeadquarterFor
+            RelationType.IsManagedBy -> SharableRelationType.IsManagedBy
         }
 
     private fun createMappingException(property: KProperty<*>, entityId: Long? = null): BpdmNullMappingException {